Move `AngularRouteTemplater` into a view helper module
